[refactor] onCreate, onRename, onDelete, onMove 관련 로직 커스텀 훅으로 숨기기
